refactor: CartItem UIModel -> Domain으로 변경
